refactored some code to move the useeffect up the component tree, created individual items page
