+++ src/components/ArticleIndividual/ArticleIndividual.jsx
@@ -1,0 +1,4 @@
+function ArticleIndividual({ currentArticleIndividual }) {
+  return <p>current article</p>;
+}
+export default ArticleIndividual;
